Bugfix JSON errors when using Celsius, Added Application Info dialog to landing screen settings. Improve Google Play update handling. Upgrade AGP to 9.2.1. Update targetSdk to 37. Update various dependencies

+++ app/src/main/java/com/weberbox/pifire/dashboard/data/model/DashDto.kt
@@ -70,7 +70,7 @@ data class DashDto(
         val title: String? = null,
         val label: String? = null,
         val eta: Int? = null,
-        val temp: Int? = null,
+        val temp: Double? = null,
         val setTemp: Int? = null,
         val maxTemp: Int? = null,
         val target: Int? = null,
